Share Insights chrome between User Testing and Swarms.
Extract flow/selection/rebuild orchestration and the fill-viewport statline shell so scenario Insights matches Swarm run detail; rename Clusters → Insights.
