+++ mcpjam-inspector/client/src/components/chatboxes/ChatboxUsagePanel.tsx
@@ -1,26 +1,24 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { MessageSquare } from "lucide-react";
-import { toast } from "@/lib/toast";
+import { MessageSquare, X } from "lucide-react";
 import type { ChatboxSettings } from "@/hooks/useChatboxes";
 import {
-  EMPTY_USAGE_FILTER,
   chipKey,
   compareThreadsForUsageList,
-  isSameSelection,
-  removeChipByKey,
-  removeChipsByKeys,
-  selectionChipsToAdd,
   threadMatchesFilterState,
-  toggleChip,
-  type InsightsSelection,
   type UsageFilterChip,
   type UsageFilterState,
 } from "@/hooks/chatbox-usage-filters";
+import {
+  useInsightsFlowController,
+  useInsightsRebuild,
+} from "@/hooks/useInsightsFlowController";
 import { useUsageInsights } from "@/hooks/useUsageInsights";
 import { ChatboxInsightsSankey } from "@/components/chatboxes/ChatboxInsightsSankey";
-import { ChatboxOutcomeCalibration } from "@/components/chatboxes/ChatboxOutcomeCalibration";
+import {
+  ChatboxOutcomeCalibration,
+  hasOutcomeFeedbackCalibration,
+} from "@/components/chatboxes/ChatboxOutcomeCalibration";
 import { ChatboxGoalOutcomeDrilldown } from "@/components/chatboxes/ChatboxGoalOutcomeDrilldown";
-import { ScrollArea } from "@mcpjam/design-system/scroll-area";
 import {
   ResizablePanelGroup,
   ResizablePanel,
@@ -29,13 +27,19 @@ import {
 import { ShareUsageThreadList } from "@/components/connection/share-usage/ShareUsageThreadList";
 import { ShareUsageThreadDetail } from "@/components/connection/share-usage/ShareUsageThreadDetail";
 import { ChatboxTopicMapPanel } from "@/components/chatboxes/ChatboxTopicMapPanel";
-import { rebuildFeedback } from "@/components/shared/usage-insights/rebuild-feedback";
-import type { ClusterTuning } from "@/lib/cluster-tuning";
+import { InsightsStatline } from "@/components/shared/usage-insights/InsightsStatline";
+import { InsightsViewToggle } from "@/components/shared/usage-insights/InsightsViewToggle";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@mcpjam/design-system/popover";
 import { buildUserTestingScenarioPath } from "@/lib/app-navigation";
 import { getShareableAppOrigin } from "@/lib/chatbox-session";
 import { usePromoteCapability } from "@/hooks/usePromoteCapability";
 import { ErrorBoundary } from "@/components/ui/error-boundary";
 import { ChatboxSessionsMetricStrip } from "@/components/chatboxes/chatbox-sessions-metric-strip";
+import { cn } from "@/lib/utils";
 
 export type ChatboxUsagePanelSection = "sessions" | "insights";
 
@@ -54,6 +58,8 @@ interface ChatboxUsagePanelProps {
    * itself (the same instance survives the insights → sessions flip).
    */
   onOpenSession?: (threadId: string) => void;
+  /** Open the Sessions tab without selecting a particular session. */
+  onOpenSessionsTab?: () => void;
 }
 
 /** Filter chip that excludes synthetic (AI-generated) sessions from the list.
@@ -67,11 +73,19 @@ const HIDE_SYNTHETIC_CHIP: UsageFilterChip = {
   label: "Hide synthetic",
 };
 
+function withHideSynthetic(filter: UsageFilterState): UsageFilterState {
+  if (filter.chips.some((c) => chipKey(c) === chipKey(HIDE_SYNTHETIC_CHIP))) {
+    return filter;
+  }
+  return { ...filter, chips: [...filter.chips, HIDE_SYNTHETIC_CHIP] };
+}
+
 export function ChatboxUsagePanel({
   chatbox,
   section,
   initialThreadId,
   onOpenSession,
+  onOpenSessionsTab,
 }: ChatboxUsagePanelProps) {
   // Scope selection to the current chatbox so switching chatboxes can't briefly
   // render a detail pane for a thread belonging to the previous chatbox.
@@ -79,27 +93,6 @@ export function ChatboxUsagePanel({
     chatboxId: string;
     threadId: string | null;
   }>({ chatboxId: chatbox.chatboxId, threadId: initialThreadId ?? null });
-  const [filter, setFilter] = useState<UsageFilterState>(EMPTY_USAGE_FILTER);
-  // Chatboxes have no synthetic traffic, so force synthetic sessions out of
-  // the fetched + rendered list regardless of the user's own filter chips
-  // (historical rows predating the removal stay in the database). Only the
-  // data path uses this; the filter UI still shows the user's chips.
-  const effectiveFilter = useMemo<UsageFilterState>(() => {
-    if (filter.chips.some((c) => chipKey(c) === chipKey(HIDE_SYNTHETIC_CHIP))) {
-      return filter;
-    }
-    return { ...filter, chips: [...filter.chips, HIDE_SYNTHETIC_CHIP] };
-  }, [filter]);
-  const [rebuildBusy, setRebuildBusy] = useState(false);
-  // Synchronous latch so double-clicks can't queue two concurrent rebuilds
-  // before React commits `rebuildBusy`.
-  const rebuildInFlightRef = useRef(false);
-  // Monotonic nonce identifying the currently-owning rebuild invocation.
-  // Each call to `handleRebuild` bumps this; the invocation captures its
-  // own value, and its `finally` only clears the latch when that nonce
-  // still matches — so a stale A-rebuild resolving after B or a later
-  // same-chatbox rebuild has started never unlocks the new one.
-  const rebuildNonceRef = useRef(0);
 
   // Promotion copies a tester's words into a durable member-owned artifact,
   // so it is member-gated server-side. Resolve the same tier here — the
@@ -119,45 +112,19 @@ export function ChatboxUsagePanel({
   const setSelectedThreadId = useCallback(
     (threadId: string | null) =>
       setSelection({ chatboxId: chatbox.chatboxId, threadId }),
-    [chatbox.chatboxId]
+    [chatbox.chatboxId],
   );
 
-  // Currently-selected flow node or link, driving the paginated drill-down.
-  // Kept next to the filter rather than derived from it: an unlabeled node's
-  // chip is a sentinel that several stages share, so the selection is not fully
-  // recoverable from the chip list alone.
-  const [flowSelection, setFlowSelection] = useState<InsightsSelection | null>(
-    null
-  );
-  // The chips the flow actually ADDED, as opposed to the ones its selection
-  // implies. They differ whenever the topic map had already written the same
-  // chip: the flow click adds nothing, and tearing down by implication would
-  // then delete the map's filter. Chip equality cannot express ownership, so it
-  // is tracked here.
-  const [flowOwnedKeys, setFlowOwnedKeys] = useState<string[]>([]);
-
-  // The breakdown backs the session flow, so the selection's own chips must NOT
-  // reach its query: they are the diagram's own OUTPUT. Feeding them back
-  // re-runs the scan filtered to the clicked node, which collapses the diagram
-  // to the single path that was selected — the selection would erase everything
-  // it was selected from. Every other dimension's chips (device, language, the
-  // forced synthetic:hide, …) still narrow it; the Sessions list, the map
-  // dimming, and the drill-down keep the full filter, because narrowing to the
-  // selection is exactly their job.
-  //
-  // Subtracted by OWNERSHIP, not by what the selection implies: the topic map
-  // and the insights strip write cluster chips too, and a chip the flow merely
-  // coincided with belongs to whoever wrote it. Removing those would throw away
-  // a community the user picked on the map and the diagram would ignore it.
-  const breakdownFilter = useMemo(
-    () => removeChipsByKeys(effectiveFilter, flowOwnedKeys),
-    [effectiveFilter, flowOwnedKeys]
-  );
+  const cohortKey = chatbox.chatboxId;
+  const flow = useInsightsFlowController({
+    cohortKey,
+    augmentFilter: withHideSynthetic,
+  });
 
   const { threads, breakdown, rebuild } = useUsageInsights({
     sourceType: "chatbox",
     sourceId: chatbox.chatboxId,
-    filters: breakdownFilter,
+    filters: flow.breakdownFilter,
     // The thread list backs Sessions; the breakdown backs the Insights grid.
     // Subscribing to each only where it renders keeps the tab flip from
     // running two scans at once. (The thread-list query takes no filters —
@@ -166,23 +133,25 @@ export function ChatboxUsagePanel({
     breakdownEnabled: section === "insights",
   });
 
+  const { rebuildBusy, handleRebuild, handleApplyTuning } = useInsightsRebuild(
+    rebuild,
+    cohortKey,
+  );
+
   // Apply filter state here (chips + preset) so chips like "Hide synthetic"
   // actually narrow the list — ShareUsageThreadList renders provided threads
   // verbatim when the panel owns the data, so filtering has to happen here.
   const sortedThreads = useMemo(() => {
     if (!threads) return undefined;
     return threads
-      .filter((t) => threadMatchesFilterState(t, effectiveFilter))
+      .filter((t) => threadMatchesFilterState(t, flow.effectiveFilter))
       .sort(compareThreadsForUsageList);
-  }, [threads, effectiveFilter]);
+  }, [threads, flow.effectiveFilter]);
 
-  // Reset below only on chatbox *switches*. Guarded by comparing against the
-  // previous chatboxId (not a mount-skip flag) so the effect is idempotent:
-  // StrictMode's dev replay re-runs it with the same chatboxId and must not
-  // wipe the deep-linked initialThreadId seed. Re-seeding from initialThreadId
-  // (instead of null) also covers the deep link's host applying after this
-  // panel mounted — that lands here as a chatbox switch, and the session param
-  // is still in the URL.
+  // Reset thread selection only on chatbox *switches*. Guarded by comparing
+  // against the previous chatboxId so StrictMode's dev replay does not wipe a
+  // deep-linked initialThreadId. Flow filter/selection reset is owned by
+  // useInsightsFlowController via cohortKey.
   const prevChatboxIdRef = useRef(chatbox.chatboxId);
   useEffect(() => {
     if (prevChatboxIdRef.current === chatbox.chatboxId) return;
@@ -191,17 +160,6 @@ export function ChatboxUsagePanel({
       chatboxId: chatbox.chatboxId,
       threadId: initialThreadId ?? null,
     });
-    setFilter(EMPTY_USAGE_FILTER);
-    // A selected flow node may name a cluster belonging to the previous chatbox.
-    setFlowSelection(null);
-    setFlowOwnedKeys([]);
-    // Reset rebuild state too — an in-flight rebuild belongs to the previous
-    // chatbox and shouldn't keep this one's button disabled. The old promise
-    // still resolves; its nonce no longer matches so its `finally` is a
-    // silent no-op.
-    rebuildNonceRef.current += 1;
-    rebuildInFlightRef.current = false;
-    setRebuildBusy(false);
   }, [chatbox.chatboxId, initialThreadId]);
 
   useEffect(() => {
@@ -231,58 +189,7 @@ export function ChatboxUsagePanel({
         threadId: sortedThreads[0]?._id ?? null,
       };
     });
-  }, [sortedThreads, chatbox.chatboxId]);
-
-  const handleToggleChip = useCallback(
-    (chip: UsageFilterChip) => setFilter((prev) => toggleChip(prev, chip)),
-    []
-  );
-  const handleClearChip = useCallback(
-    (key: string) => setFilter((prev) => removeChipByKey(prev, key)),
-    []
-  );
-
-  // Flow node / link click. Uses applySelection (not a series of toggleChip
-  // calls) so the themes are REPLACED together — chips OR within an axis, so
-  // toggling a second theme on one axis would widen rather than narrow.
-  //
-  // The previous selection is subtracted BY IDENTITY rather than by clearing
-  // the axis: the diagram is not the only writer of theme chips — the topic map
-  // writes a goal chip when a community is clicked — so clearing by axis would
-  // silently drop a filter the user set elsewhere and widen the cohort behind
-  // their back.
-  //
-  // `flowSelection` is the single source of truth for what is open, and the
-  // reopen-vs-close decision is made once, here, from it. Both setters are
-  // called at the top level (never one nested inside the other's updater) so
-  // StrictMode's double-invoked reducers cannot toggle the filter twice.
-  const handleSelectFlow = useCallback(
-    (next: InsightsSelection) => {
-      const isAlreadyOpen = isSameSelection(flowSelection, next);
-      // Computed from the current filter rather than inside a state updater, so
-      // StrictMode's double-invoked reducers cannot record ownership twice.
-      const cleared = removeChipsByKeys(filter, flowOwnedKeys);
-      if (isAlreadyOpen) {
-        setFilter(cleared);
-        setFlowSelection(null);
-        setFlowOwnedKeys([]);
-        return;
-      }
-      const added = selectionChipsToAdd(cleared, next);
-      setFilter({ ...cleared, chips: [...cleared.chips, ...added] });
-      setFlowSelection(next);
-      setFlowOwnedKeys(added.map(chipKey));
-    },
-    [filter, flowSelection, flowOwnedKeys]
-  );
-
-  const handleCloseFlow = useCallback(() => {
-    // Drops only what the flow put there, so the diagram cannot keep a theme
-    // highlighted with nothing open behind it and an unrelated filter survives.
-    setFilter((prev) => removeChipsByKeys(prev, flowOwnedKeys));
-    setFlowSelection(null);
-    setFlowOwnedKeys([]);
-  }, [flowOwnedKeys]);
+  }, [sortedThreads, chatbox.chatboxId, setSelectedThreadId]);
 
   // Topic-map dot click → open that session in the Sessions tab. Clear the
   // filter so an active cluster chip can't hide the target thread (the
@@ -290,87 +197,160 @@ export function ChatboxUsagePanel({
   const handleOpenSessionFromMap = useCallback(
     (sessionId: string) => {
       setSelection({ chatboxId: chatbox.chatboxId, threadId: sessionId });
-      setFilter(EMPTY_USAGE_FILTER);
-      // The cleared filter no longer encodes a selection, so the diagram
-      // highlight and the drill-down panel have to go with it.
-      setFlowSelection(null);
-      setFlowOwnedKeys([]);
+      flow.clearAllFilters();
       onOpenSession?.(sessionId);
     },
-    [chatbox.chatboxId, onOpenSession]
-  );
-
-  const handleRebuild = useCallback(
-    async (args?: { tuning?: ClusterTuning; force?: boolean }) => {
-      if (rebuildInFlightRef.current) return;
-      // Bump the nonce and capture it. The `finally` compares against the
-      // current nonce: any later invocation (A→B→A→rebuild-again, or just
-      // same-chatbox trigger-again after a chatbox-switch reset) bumps the
-      // counter, so the earlier promise's `finally` finds a mismatch and
-      // leaves the latch alone for the live rebuild.
-      rebuildNonceRef.current += 1;
-      const myNonce = rebuildNonceRef.current;
-      rebuildInFlightRef.current = true;
-      setRebuildBusy(true);
-      try {
-        const result = await rebuild(args);
-        const { tone, message } = rebuildFeedback(result);
-        toast[tone](message);
-      } catch (error) {
-        toast.error(
-          error instanceof Error
-            ? error.message
-            : "Rebuild failed. Try again in a few minutes."
-        );
-      } finally {
-        if (rebuildNonceRef.current === myNonce) {
-          rebuildInFlightRef.current = false;
-          setRebuildBusy(false);
-        }
-      }
-    },
-    [rebuild]
-  );
-
-  const handleApplyTuning = useCallback(
-    (tuning: ClusterTuning, opts?: { force?: boolean }) => {
-      void handleRebuild({ tuning, ...opts });
-    },
-    [handleRebuild]
+    [chatbox.chatboxId, flow.clearAllFilters, onOpenSession],
   );
 
   if (section === "insights") {
+    const viewToggle = (
+      <InsightsViewToggle
+        view={flow.view}
+        onChange={flow.setView}
+        testId="chatbox-insights-view-toggle"
+      />
+    );
+
+    const chipRow =
+      flow.dismissibleChips.length > 0 ? (
+        <div className="flex flex-wrap items-center gap-1.5 px-5 py-2">
+          {flow.dismissibleChips.map((chip) => {
+            const key = chipKey(chip);
+            const label =
+              chip.kind === "cluster"
+                ? (chip.label ?? "Cluster")
+                : (chip.label ?? `${chip.key}: ${chip.value}`);
+            return (
+              <button
+                key={key}
+                type="button"
+                onClick={() => flow.handleClearChip(key)}
+                className="inline-flex items-center gap-1 rounded-full border bg-muted/50 px-2 py-0.5 text-xs hover:bg-muted"
+              >
+                <span>{label}</span>
+                <X className="size-3" />
+              </button>
+            );
+          })}
+        </div>
+      ) : null;
+
+    const selectionOpen = flow.flowSelection !== null;
+    const scope = {
+      kind: "chatbox" as const,
+      chatboxId: chatbox.chatboxId,
+    };
+    const showFeedbackChip = hasOutcomeFeedbackCalibration(breakdown);
+
     return (
-      <div className="flex h-full min-h-0 flex-col overflow-hidden">
-        <ScrollArea className="max-h-[62%] shrink-0 border-b">
-          <ChatboxInsightsSankey
-            breakdown={breakdown}
-            selection={flowSelection}
-            onSelectNode={handleSelectFlow}
-            onSelectLink={handleSelectFlow}
-            onRebuild={handleRebuild}
-            rebuildBusy={rebuildBusy}
-            onApplyTuning={handleApplyTuning}
-          />
-          <ChatboxGoalOutcomeDrilldown
-            scope={{ kind: "chatbox", chatboxId: chatbox.chatboxId }}
-            selection={flowSelection}
-            filter={effectiveFilter}
-            onClose={handleCloseFlow}
-            onOpenSession={handleOpenSessionFromMap}
-          />
-          <ChatboxOutcomeCalibration breakdown={breakdown} />
-        </ScrollArea>
-        <div className="min-h-0 flex-1">
-          <ChatboxTopicMapPanel
-            chatboxId={chatbox.chatboxId}
-            filter={filter}
-            onToggleChip={handleToggleChip}
-            onClearChip={handleClearChip}
-            onRebuild={handleRebuild}
-            rebuildBusy={rebuildBusy}
-            onOpenSession={handleOpenSessionFromMap}
-          />
+      <div
+        className="flex h-full min-h-0 flex-col gap-2 overflow-hidden px-8 py-4"
+        data-testid="chatbox-insights-panel"
+      >
+        <InsightsStatline
+          breakdown={breakdown}
+          filter={flow.filter}
+          flowSelection={flow.flowSelection}
+          onSelectFlow={flow.handleSelectFlow}
+          onToggleChip={flow.handleToggleChip}
+          onOpenSessionsTab={onOpenSessionsTab}
+          strugglesSlot={
+            showFeedbackChip ? (
+              <Popover>
+                <PopoverTrigger asChild>
+                  <button
+                    type="button"
+                    className="inline-flex min-w-0 items-center gap-1 rounded-md border border-border/50 bg-muted/25 px-2 py-0.5 text-xs font-medium tabular-nums transition-colors hover:bg-muted/50"
+                    data-testid="chatbox-insights-feedback-chip"
+                  >
+                    Feedback
+                  </button>
+                </PopoverTrigger>
+                <PopoverContent
+                  align="start"
+                  className="w-[28rem] max-w-[90vw] p-0"
+                >
+                  <div className="flex max-h-[60vh] min-h-0 flex-col overflow-y-auto">
+                    <ChatboxOutcomeCalibration breakdown={breakdown} />
+                  </div>
+                </PopoverContent>
+              </Popover>
+            ) : null
+          }
+          trailing={viewToggle}
+          testId="chatbox-insights-statline"
+        />
+        <div className="relative flex min-h-0 flex-1 overflow-hidden">
+          <div className="flex min-w-0 flex-1 flex-col overflow-hidden">
+            {flow.view === "clusters" ? (
+              <div className="flex h-full min-h-0 flex-col">
+                {chipRow}
+                <div className="min-h-0 flex-1">
+                  <ChatboxTopicMapPanel
+                    chatboxId={chatbox.chatboxId}
+                    filter={flow.filter}
+                    onToggleChip={flow.handleToggleChip}
+                    onClearChip={flow.handleClearChip}
+                    onRebuild={handleRebuild}
+                    rebuildBusy={rebuildBusy}
+                    onOpenSession={handleOpenSessionFromMap}
+                  />
+                </div>
+              </div>
+            ) : (
+              <div className="flex h-full min-h-0 flex-col overflow-hidden">
+                <div className="min-h-0 flex-1 overflow-hidden">
+                  <ChatboxInsightsSankey
+                    breakdown={breakdown}
+                    selection={flow.flowSelection}
+                    onSelectNode={flow.handleSelectFlow}
+                    onSelectLink={flow.handleSelectFlow}
+                    onRebuild={handleRebuild}
+                    rebuildBusy={rebuildBusy}
+                    onApplyTuning={handleApplyTuning}
+                    showLinkThreshold
+                    fillHeight
+                  />
+                </div>
+                {chipRow}
+              </div>
+            )}
+          </div>
+          {flow.view === "flow" ? (
+            <div
+              className={cn(
+                selectionOpen
+                  ? "absolute inset-0 z-10 bg-background sm:static sm:w-[22rem] lg:w-[24rem] sm:shrink-0 sm:border-l sm:border-border/40"
+                  : "hidden",
+              )}
+              data-testid="chatbox-insights-drill-panel"
+              aria-hidden={!selectionOpen}
+            >
+              {/* Always mounted (hidden when closed) so close toggles
+                  `enabled: false` instead of unmounting — the flow-selection
+                  tests pin that contract. */}
+              <ChatboxGoalOutcomeDrilldown
+                scope={scope}
+                selection={flow.flowSelection}
+                filter={flow.effectiveFilter}
+                variant="panel"
+                onClose={flow.handleCloseFlow}
+                onOpenSession={handleOpenSessionFromMap}
+                footer={
+                  onOpenSessionsTab ? (
+                    <button
+                      type="button"
+                      className="self-start text-xs font-medium text-primary hover:underline"
+                      onClick={onOpenSessionsTab}
+                    >
+                      Open in Sessions tab →
+                    </button>
+                  ) : null
+                }
+              />
+            </div>
+          ) : null}
         </div>
       </div>
     );
@@ -399,7 +379,7 @@ export function ChatboxUsagePanel({
                   threads={sortedThreads}
                   selectedThreadId={selectedThreadId}
                   onSelectThread={setSelectedThreadId}
-                  filterState={effectiveFilter}
+                  filterState={flow.effectiveFilter}
                 />
               </div>
             </div>
@@ -412,7 +392,7 @@ export function ChatboxUsagePanel({
                   threadId={selectedThreadId}
                   sessionLink={`${getShareableAppOrigin()}${buildUserTestingScenarioPath(
                     chatbox.chatboxId,
-                    { session: selectedThreadId }
+                    { session: selectedThreadId },
                   )}`}
                   promote={
                     chatbox.projectId
